Returned type of variable of method getCardBalance() in DashboardPage changed from integer to double

+++ src/test/java/ru/netology/page/DashboardPage.java
@@ -31,7 +31,7 @@ public class DashboardPage {
         return new TransactionPage();
     }
 
-    public int getCardBalance(String id) {
+    public double getCardBalance(String id) {
         // перебрать все карты и найти по атрибуту data-test-id
         val text = cards.findBy(attribute("data-test-id", id)).text();
         return extractBalance(text);
